Update to current modules and remove problematic test (#122)
This is the first change to stabilize things here. The
container_structure_test has [never supported multi-arch
images](https://github.com/GoogleContainerTools/container-structure-test/issues/435).
It only passed because these example cases didn't end up building
multi-arch images. We will address this in a followup PR.

Moving to the last v1 `oci_rules` release (v1.8.0) gives us better
insight into failure cases. The current v2 releases change a lot of
behavior which we will look at in a followup PR.

The rest of the module updates are benign and bring us up to a more
modern state. We also run
[buildifier](https://github.com/bazelbuild/buildtools/blob/main/buildifier/README.md)
to cleanup the BUILD files.

Bump apko to v0.27.1 which isn't the latest- but is the latest that
works with this current code base. We will address this in a followup PR
as well.