Removes deprecated toolchain_transition flag (#246)
Removes the `incompatible_use_toolchain_transition` attribute from the
resolved toolchain rule. This attribute was deprecated and turned into a
silent no-op as far back as bazel 6.5
https://bazel.build/versions/6.5.0/rules/lib/globals
In bazel 9, it considers its inclusion a syntax error.
This cleans up the attribute for good hygiene, and to help prepare the
ruleset for bazel 9.